requirements.txt
datasets>=2.14.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
torch>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
